Creating library for compiling Solidity contracts
